Correct datatype for mjd

+++ filhead.c
@@ -32,7 +32,7 @@ void send_long(FILE *outfile, char *name, long number){
     fwrite(&number, sizeof(long), 1, outfile);
 }
 
-void filterbank_header(FILE *outfile, char *infilename, char *jname, int mjd, double freq, double bw, int nchan, double tsmpl){
+void filterbank_header(FILE *outfile, char *infilename, char *jname, double mjd, double freq, double bw, int nchan, double tsmpl){
     
     int nbits = 16;
     int nbeams = 1;
